refactor: split sosotalk post detail page responsibilities
