+++ src/widgets/sosotalk/ui/sosotalk-post-detail/model/index.ts
@@ -1,0 +1,8 @@
+export {
+  SOSOTALK_AUTHOR_IMAGE_FALLBACK,
+  formatSosoTalkRelativeTime,
+  mapCommentToCommentItemData,
+} from './sosotalk-post-detail-page.utils';
+export { useSosoTalkPostDetailCommentActions } from './use-sosotalk-post-detail-comment-actions';
+export { useSosoTalkPostDetailPage } from './use-sosotalk-post-detail-page';
+export { useSosoTalkPostDetailPostActions } from './use-sosotalk-post-detail-post-actions';
